refactor(compiler): move LSP Position/Range conversion to separate file
