+++ new-new-packages/compiler/src/language_server/open_file_manager.rs
@@ -1,8 +1,9 @@
 use im::HashMap;
 use lsp_types::{
-    DidChangeTextDocumentParams, DidCloseTextDocumentParams, DidOpenTextDocumentParams, Position,
-    Url,
+    DidChangeTextDocumentParams, DidCloseTextDocumentParams, DidOpenTextDocumentParams, Url,
 };
+
+use crate::language_server::utils::RangeToUtf8ByteOffset;
 
 #[derive(Debug)]
 pub struct OpenFileManager {
@@ -38,9 +39,13 @@ impl OpenFileManager {
                 for change in content_changes {
                     match change.range {
                         Some(range) => {
-                            let start = Self::position_to_utf8_byte_offset(text, &range.start);
-                            let end = Self::position_to_utf8_byte_offset(text, &range.end);
-                            *text = format!("{}{}{}", &text[..start], &change.text, &text[end..]);
+                            let range = range.to_utf8_byte_offset(text);
+                            *text = format!(
+                                "{}{}{}",
+                                &text[..range.start],
+                                &change.text,
+                                &text[range.end..]
+                            );
                         }
                         None => *text = change.text,
                     }
@@ -48,43 +53,6 @@ impl OpenFileManager {
                 log::info!("New text: {:?}", text);
             })
             .or_insert_with(|| panic!("Received a change for a file that was not opened."));
-    }
-
-    fn position_to_utf8_byte_offset(text: &str, position: &Position) -> usize {
-        let mut line_index = 0;
-        let mut line_offset = 0;
-        while line_index < position.line {
-            match text.bytes().nth(line_offset).unwrap() {
-                b'\n' => {
-                    line_index += 1;
-                    line_offset += 1;
-                }
-                _ => {
-                    line_offset += 1;
-                }
-            }
-        }
-
-        let mut line_length_bytes = 0;
-        loop {
-            match text.bytes().nth(line_offset + line_length_bytes) {
-                Some(b'\r' | b'\n') | None => break,
-                Some(_) => line_length_bytes += 1,
-            }
-        }
-
-        let line = &text[line_offset..line_offset + line_length_bytes];
-
-        let words = line.encode_utf16().collect::<Vec<_>>();
-        let char_offset = if position.character as usize >= words.len() {
-            line_length_bytes
-        } else {
-            String::from_utf16(&words[0..position.character as usize])
-                .unwrap()
-                .len()
-        };
-
-        line_offset + char_offset
     }
 
     pub async fn did_close(&mut self, params: DidCloseTextDocumentParams) {
